Moved class declaration of FaceRecognitionUI to FaceRecognitionUI.h
