+++ RealTimeFaceMaskDetector/TCPClientUI/TCPClient.h
@@ -13,6 +13,7 @@
 #include <QImage>
 #include <QDebug>
 
+#include "FaceRecognitionUI.h"
 #include "ui_FaceRecognitionUI.h"
 #include "MaskRecognizer.h"
 
@@ -39,25 +40,3 @@ private:
     SOCKET m_socket;
     size_t m_size;    
 };
-
-class FaceRecognitionUI : public QWidget
-{
-    Q_OBJECT
-
-public:
-    FaceRecognitionUI(QWidget* parent = Q_NULLPTR);
-    ~FaceRecognitionUI();
-
-    bool m_exit_button_clicked = false;
-    void Recognize(TCPClient& client);
-    QImage Mat2QImage(cv::Mat const& src);
-    friend std::vector<char> GetImageBytesVector();
-
-public slots:
-    void onExitButtonClicked();
-
-public:
-    Ui::FaceRecognitionUI ui;
-};
-
-extern std::unique_ptr<MaskRecognizer> m_mask_recognizer;
